requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.8.6
python-dotenv>=1.0.1
aiohttp>=3.9.5
openai>=1.37.1
colorlog>=6.8.2